requirements.txt
flask==3.1.0
gunicorn==23.0.0
chromadb==1.5.5
sentence-transformers==3.4.1
groq==0.13.1
python-dotenv==1.1.0
pypdf==5.4.0
markdown-it-py==3.0.0
beautifulsoup4==4.13.3
lxml==5.3.1
numpy==2.2.4
